added server and started user test. Added rake capy_test.
