+++ Rakefile
@@ -8,6 +8,10 @@ task :test do
   Dir.glob('./test/**/*_test.rb') { |file| require file }
 end
 
+task :capy_test do
+  Dir.glob('./test/features/*_test.rb') { |file| require file }
+end
+
 
 namespace :sanitation do
   desc "Check line lengths & whitespace with Cane"
