+++ test/features/user_sees_aggregate_site_data_test.rb
@@ -1,12 +1,30 @@
 require './test/test_helper'
 class AggregateDataTest < FeatureTest
-# 
-#   def test_user_can_see_most_requested_to_least_requested_urls
-#     As a user When I visit "sources/identifier"
-#     visit '/sources/jumpstartlab'
-#     assert_equal '/sources/jumpstartlab', current_path
-#     assert page.has_content?('')
-# Then I should see most requested to least requested URLs
-#   end
 
+  def app
+    TrafficSpy::Server
   end
+
+  def setup
+    DatabaseCleaner.start
+    TestData.clients.each do |client|
+      ClientValidator.validate(client)
+    end
+    TestData.payloads.each do |payload|
+      client = Client.all.first.identifier
+      PayloadValidator.validate(payload["payload"], client)
+    end
+  end
+
+  def teardown
+    DatabaseCleaner.clean
+  end
+
+  def test_user_can_see_most_requested_to_least_requested_urls
+    visit '/sources/jumpstartlab'
+    assert_equal '/sources/jumpstartlab', current_path
+    save_and_open_page
+    assert page.has_content?('jumpstartlab.com/blog')
+  end
+
+end
